% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartiteNetworkWidget.R
\name{bipartiteNetworkOutput}
\alias{bipartiteNetworkOutput}
\title{Shiny bindings for bipartiteNetwork}
\usage{
bipartiteNetworkOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'}, '400px', or 'auto')}
}
\description{
Output function for using bipartiteNetwork within Shiny apps
}
