% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpg_all_cont}
\alias{mpg_all_cont}
\title{\code{mpg} continuous variables}
\format{
\subsection{\code{mpg_all_cont}}{

A data frame with 234 rows and 5 columns:
\describe{
\item{cty}{City miles per gallon}
\item{hwy}{Highway miles per gallon}
\item{year}{Year of manufacture}
\item{cyl}{Number of cylinders}
\item{displ}{Displacement in litres}
}
}
}
\source{
\url{https://github.com/tidyverse/ggplot2}
}
\usage{
mpg_all_cont
}
\description{
A subset of the \code{mpg} dataset from the \code{ggplot2} package.
It contains only the integer and numeric variables.
}
\keyword{datasets}
