% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpg_cty_hwy}
\alias{mpg_cty_hwy}
\title{\code{mpg} city and highway miles per gallon}
\format{
\subsection{\code{mpg_cty_hwy}}{

A data frame with 234 rows and 2 columns:
\describe{
\item{cty}{City miles per gallon}
\item{hwy}{Highway miles per gallon}
}
}
}
\source{
\url{https://github.com/tidyverse/ggplot2}
}
\usage{
mpg_cty_hwy
}
\description{
A subset of the \code{mpg} dataset from the \code{ggplot2} package.
It contains only the \code{cty} and \code{hwy} variables.
}
\keyword{datasets}
