% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mpg_year_cyl}
\alias{mpg_year_cyl}
\title{\code{mpg} year, cylinders, and displacement}
\format{
\subsection{\code{mpg_year_cyl}}{

A data frame with 234 rows and 3 columns:
\describe{
\item{year}{Year of manufacture}
\item{cyl}{Number of cylinders}
\item{displ}{Displacement in litres}
}
}
}
\source{
\url{https://github.com/tidyverse/ggplot2}
}
\usage{
mpg_year_cyl
}
\description{
A subset of the \code{mpg} dataset from the \code{ggplot2} package.
It contains only the \code{year}, \code{cyl} and \code{displ} variables.
}
\keyword{datasets}
