% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pathways}
\alias{pathways}
\title{BONUS-CF Pathway Abundance}
\format{
\subsection{\code{pathways}}{

A data frame with 1279 rows and 496 columns:
\describe{
\item{X1CMET2.PWY..N10.formyl.tetrahydrofolate.biosynthesis..EUPATH_0009248_1CMET2_PWY.}{Abundance of pathway with BioCyc ID 1CMET2.PWY.}
...
}
}
}
\source{
\url{https://microbiomedb.org}
}
\usage{
pathways
}
\description{
Data from MicrobiomeDB.org, study BONUS-CF
}
\keyword{datasets}
