% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartiteNetworkWidget.R
\name{renderBipartiteNetwork}
\alias{renderBipartiteNetwork}
\title{Shiny bindings for bipartiteNetwork}
\usage{
renderBipartiteNetwork(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{an expression that generates bipartiteNetwork}

\item{env}{the environment in which to evaluate \code{expr}}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable}
}
\description{
Render function for using bipartiteNetwork within Shiny apps
}
