% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unipartiteNetwork.R
\name{renderUnipartiteNetwork}
\alias{renderUnipartiteNetwork}
\title{Shiny bindings for unipartiteNetwork}
\usage{
renderUnipartiteNetwork(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{an expression that generates unipartiteNetwork}

\item{env}{the environment in which to evaluate \code{expr}}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable}
}
\description{
Render function for using unipartiteNetwork within Shiny apps
}
