% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{BONUS-CF Relative Abundance of Species}
\format{
\subsection{\code{species}}{

A data frame with 1279 rows and 602 columns:
\describe{
\item{Oscillibacter.sp..57_20..EUPATH_0009257_Bacteria_0.}{Relative Abundance of Oscillibacter sp. 57_20}
...
}
}
}
\source{
\url{https://microbiomedb.org}
}
\usage{
species
}
\description{
Data from MicrobiomeDB.org, study BONUS-CF
}
\keyword{datasets}
