% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unipartiteNetwork.R
\name{unipartiteNetwork}
\alias{unipartiteNetwork}
\title{unipartiteNetwork}
\usage{
unipartiteNetwork(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{a data frame with source, target, and value columns}

\item{width}{the width of the widget}

\item{height}{the height of the widget}

\item{elementId}{the ID of the widget}
}
\description{
Create a unipartiteNetwork widget
}
