% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unipartiteNetwork.R
\name{unipartiteNetworkOutput}
\alias{unipartiteNetworkOutput}
\title{Shiny bindings for unipartiteNetwork}
\usage{
unipartiteNetworkOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'}, '400px', or 'auto')}
}
\description{
Output function for using unipartiteNetwork within Shiny apps
}
